fix(richTextEditor): modify RestoreFromLocalStoragePlugin -> RestoreContentFromData
